use pydantic for task state updates
